requirements.txt
click==7.1.2
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dotenv==0.17.0
Werkzeug==1.0.1
psycopg2==2.8.4
Flask-Migrate==2.5.2
Flask-Script==2.0.6